requirements.txt
Django==5.2.7
djangorestframework
djangorestframework-simplejwt
django-postgres-extra
django-cors-headers
psycopg2-binary
channels
daphne
pillow
python-decouple
python-dotenv
supabase
